refactoring filters to filters.py. Added ProductFilter.
