+++ openapi.yaml
@@ -72,6 +72,18 @@ paths:
     get:
       operationId: api_v1_all_products_list
       parameters:
+      - in: query
+        name: category
+        schema:
+          type: integer
+      - in: query
+        name: collection
+        schema:
+          type: integer
+      - in: query
+        name: is_ukrainian_cashback
+        schema:
+          type: boolean
       - name: limit
         required: false
         in: query
@@ -82,6 +94,14 @@ paths:
         required: false
         in: query
         description: The initial index from which to return the results.
+        schema:
+          type: integer
+      - in: query
+        name: subcategory
+        schema:
+          type: integer
+      - in: query
+        name: year_collection
         schema:
           type: integer
       tags:
@@ -1049,6 +1069,18 @@ paths:
       operationId: api_v1_product_list
       description: CRUD для продуктів
       parameters:
+      - in: query
+        name: category
+        schema:
+          type: integer
+      - in: query
+        name: collection
+        schema:
+          type: integer
+      - in: query
+        name: is_ukrainian_cashback
+        schema:
+          type: boolean
       - name: limit
         required: false
         in: query
@@ -1059,6 +1091,14 @@ paths:
         required: false
         in: query
         description: The initial index from which to return the results.
+        schema:
+          type: integer
+      - in: query
+        name: subcategory
+        schema:
+          type: integer
+      - in: query
+        name: year_collection
         schema:
           type: integer
       tags:
